Move branding env variable to pkg/branding
