+++ pkg/branding/env.go
@@ -1,14 +1,14 @@
-package scheduler
+package branding
 
 import "github.com/stackrox/rox/pkg/env"
 
 const (
 	// ProductBranding should hold RHACS_BRANDING or STACKROX_BRANDING
-	productBrandingEnvName = "ROX_PRODUCT_BRANDING"
+	ProductBrandingEnvName = "ROX_PRODUCT_BRANDING"
 )
 
 var (
-	productBrandingSetting = env.RegisterSetting(productBrandingEnvName)
+	productBrandingSetting = env.RegisterSetting(ProductBrandingEnvName, env.WithDefault("RHACS_BRANDING"))
 )
 
 // GetBrandedProductName returns the environment variable ROX_BRANDING_NAME value
